+++ packages/twenty-ui/src/testing/index.ts
@@ -15,12 +15,5 @@ export type {
 } from './decorators/CatalogDecorator';
 export { CatalogDecorator } from './decorators/CatalogDecorator';
 export { ComponentDecorator } from './decorators/ComponentDecorator';
-export type { RouteParams } from './decorators/ComponentWithRouterDecorator';
-export {
-  isRouteParams,
-  computeLocation,
-  ComponentWithRouterDecorator,
-} from './decorators/ComponentWithRouterDecorator';
-export { RouterDecorator } from './decorators/RouterDecorator';
 export { AVATAR_URL_MOCK } from './mocks/avatarUrlMock';
 export type { CatalogStory } from './types/CatalogStory';